use default inspect impl
